api/requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
python-multipart
pytesseract
transformers
protobuf